Fix install.sh
Check for existence before backing up and remove old files.

+++ install.sh
@@ -2,10 +2,12 @@
 
 script_dir=$(dirname "$(readlink -f "$0")")
 
-mv "$HOME/.bash_profile" "$HOME/.bash_profile_backup"
-mv "$HOME/.bashrc" "$HOME/.bashrc_backup"
 
+[ -f "$HOME/.bash_profile" ] && mv "$HOME/.bash_profile" "$HOME/.bash_profile_backup"
+[ -f "$HOME/.bashrc" ] && mv "$HOME/.bashrc" "$HOME/.bashrc_backup"
 
+rm -f $HOME/.bash_profile
+rm -f $HOME/.bashrc
 # Create symlinks
 ln -s "$script_dir/bash_profile" "$HOME/.bash_profile"
 ln -s "$script_dir/bashrc" "$HOME/.bashrc"
